perf: hoist the wide dense route's activation conversion to one pass per launch
- The wide Q8_0 dense prefill route converted its activation inside the K
  loop, once per column block, and re-read the f32 activation the same number
  of times. The measured composite (conversion pass + f16-input tile, no
  sharing) is 553.9 ms over the route's 1056 launches in a code-p4096 prefill
  with bit-identical output; the packet's 590.6 ms assumed the conversion was
  shared across consumers, which is worth only 7.5 ms.
- Wired at launch time: a bounded runner-owned workspace (shared with the B2
  staging allocation, single-launch-scoped on one stream), one conversion per
  launch, dispatch to the registered f16-in sibling, f32-input owner kept as
  the fallback. Profile-scoped with HIPENGINE_GGUF_Q8_DENSE_WIDE_F16_ACT=0 as
  the opt-out.
- The first census showed the swap inert at 792 launches: the sibling key is
  registered under hip_gfx1100 while the dispatch backend is the hip_gfx1151
  package, so the launch-time lookup needed the same package population the
  dispatch does. Pinned by a test that drives launch_gguf_linear rather than
  the helper, which is the level the bug was at.
- GPU tests pin bit identity on the f32-in, hoisted and fallback paths
  including a 257-row tail and a 2112-column tail, per-launch conversion
  rather than a cache, and a sentinel guard on the staged region.
- The synthetic Q8_0/Q5_K/Q6_K block builders overflowed int16/uint8 for the
  production column counts (make_q8_0_weight(10240, 2560) could not build);
  widened to int32 with every in-range value unchanged.
- Engine measurement is blocked on host memory: a running hipengine serve
  process holds 100.9 GB of GTT and the model needs 95.9 GB. No saving is
  claimed yet; the census and route A/B are scripted and ready.

+++ tests/test_gpu_gguf_k_gemv.py
@@ -93,7 +93,7 @@ def make_q6_k_weight(out_features: int, in_features: int) -> np.ndarray:
 
 def _make_q8_0_block(out_idx: int, block_idx: int) -> np.ndarray:
     d = np.float16(0.03125 * (1 + (out_idx % 5)))
-    q = ((np.arange(32, dtype=np.int16) + out_idx * 7 + block_idx * 3) % 31 - 15).astype(
+    q = ((np.arange(32, dtype=np.int32) + out_idx * 7 + block_idx * 3) % 31 - 15).astype(
         np.int8
     )
     return np.concatenate([np.asarray([d], dtype=np.float16).view(np.uint8), q.view(np.uint8)])
@@ -102,10 +102,10 @@ def _make_q8_0_block(out_idx: int, block_idx: int) -> np.ndarray:
 def _make_q5_k_block(out_idx: int, block_idx: int) -> np.ndarray:
     d = np.float16(0.015625 * (1 + (out_idx % 5)))
     dmin = np.float16(0.0078125 * (1 + (block_idx % 3)))
-    scales = ((np.arange(8, dtype=np.uint8) * 3 + out_idx + block_idx) % 63 + 1).astype(
+    scales = ((np.arange(8, dtype=np.int32) * 3 + out_idx + block_idx) % 63 + 1).astype(
         np.uint8
     )
-    mins = ((np.arange(8, dtype=np.uint8) * 5 + 2 * out_idx + block_idx) % 17).astype(
+    mins = ((np.arange(8, dtype=np.int32) * 5 + 2 * out_idx + block_idx) % 17).astype(
         np.uint8
     )
     q = ((np.arange(QK_K, dtype=np.uint16) + out_idx * 7 + block_idx * 11) % 32).astype(
@@ -132,10 +132,10 @@ def _make_q5_k_block(out_idx: int, block_idx: int) -> np.ndarray:
 
 def _make_q6_k_block(out_idx: int, block_idx: int) -> np.ndarray:
     d = np.float16(0.0107421875 * (1 + (out_idx % 3)))
-    scales = ((np.arange(16, dtype=np.int16) * 3 + out_idx - block_idx) % 31 - 15).astype(
+    scales = ((np.arange(16, dtype=np.int32) * 3 + out_idx - block_idx) % 31 - 15).astype(
         np.int8
     )
-    q_signed = ((np.arange(QK_K, dtype=np.int16) + out_idx * 5 + block_idx * 9) % 64 - 32)
+    q_signed = ((np.arange(QK_K, dtype=np.int32) + out_idx * 5 + block_idx * 9) % 64 - 32)
     q = (q_signed + 32).astype(np.uint8)
     ql = np.zeros(128, dtype=np.uint8)
     qh = np.zeros(64, dtype=np.uint8)
